Same pipes
This update makes it so that the pipes are all the same size. In the near future I will make it so that the pipes come from a constant array of pipes so that there is some variation in height but not to much for our bot. That way its memory from run to run matters.